feat(pdf): InvoiceTemplate composer replaces hardcoded template (FU-035-H1)
- Extract buildSectionData to lib/pdf/build-invoice-section-data.js (pure,
  JSX-free, test-friendly) mirroring lib/pdf/render-delivery-order.js pattern
- Update tests/invoice-build-section-data.test.mjs import to new lib path
- Replace components/pdf/InvoiceTemplate.js stub with full section-registry
  composer: iterates getSectionsForDocumentType('invoice'), calls
  computeVisibility / extractColors, dispatches to Header / InvoiceDetails /
  AddressDetails / OrderDetails / CommodityDetails / ChargeDetails / Notes /
  Disclaimer / MoveBlock / DocumentFooter
- address_details renderSection applies bill_to→customer field-ID translation
  and customerLabel:'Bill To' for AddressDetails shared component
- Re-exports buildSectionData from InvoiceTemplate.js for back-compat

Tests: 225/226 pass (1 pre-existing fire-trigger-entity-aware failure)

Resolves: FU-035-H1 Task 12
